update to reduce listing

+++ apis/marketplace.js
@@ -65,7 +65,7 @@ const isOfferCancelNotifiable = async (receiver, nft, tokenID) => {
 router.post("/itemListed", service_auth, async (req, res) => {
   try {
     const { args, blockNumber, transactionHash } = req.body;
-    const [ ownerC, nftC, tokenIdBN, quantityBN, paytokenC, pricePerItemBN, startingTimeBN ] = args;
+    const [ownerC, nftC, tokenIdBN, quantityBN, paytokenC, pricePerItemBN, startingTimeBN] = args;
 
     console.log('args', args);
 
@@ -139,12 +139,12 @@ router.post("/itemListed", service_auth, async (req, res) => {
         pricePerItem,
         itemPayToken.address
       );
-    } catch(err) {
+    } catch (err) {
       Logger.error("[ItemListed] Failed to notify followers: ", err.message);
     }
 
     Logger.info("[ItemListed] Success: ", { transactionHash, blockNumber });
-    return res.json({status: "success"});
+    return res.json({ status: "success" });
   } catch (error) {
     Logger.info("[ItemListed] Failed: ", { transactionHash: req.body.transactionHash, blockNumber: req.body.blockNumber });
     Logger.error(error);
@@ -156,7 +156,7 @@ router.post("/itemListed", service_auth, async (req, res) => {
 router.post("/itemSold", service_auth, async (req, res) => {
   try {
     const { args, blockNumber, transactionHash } = req.body;
-    const [ sellerC, buyerC, nftC, tokenIdBN, quantityBN, paytokenC, unitPriceBN, pricePerItemBN ] = args;
+    const [sellerC, buyerC, nftC, tokenIdBN, quantityBN, paytokenC, unitPriceBN, pricePerItemBN] = args;
 
     const seller = sellerC.toLowerCase();
     const buyer = buyerC.toLowerCase();
@@ -192,9 +192,9 @@ router.post("/itemSold", service_auth, async (req, res) => {
 
       try {
         // send mail here to buyer first
-        let account = await Account.findOne({address: buyer});
+        let account = await Account.findOne({ address: buyer });
         // checks if user listens
-        let ns = await NotificationSetting.findOne({address: buyer});
+        let ns = await NotificationSetting.findOne({ address: buyer });
         if (account && ns.sNftBuy) {
           let to = account.email;
           let alias = account.alias;
@@ -223,8 +223,8 @@ router.post("/itemSold", service_auth, async (req, res) => {
 
       try {
         // checks if user listens
-        const account = await Account.findOne({address: seller});
-        const ns = await NotificationSetting.findOne({address: seller});
+        const account = await Account.findOne({ address: seller });
+        const ns = await NotificationSetting.findOne({ address: seller });
         if (account && ns.sNftSell) {
           let to = account.email;
           let alias = account.alias;
@@ -252,7 +252,7 @@ router.post("/itemSold", service_auth, async (req, res) => {
       }
     }
 
-      // add new trade history
+    // add new trade history
     try {
       const existingHistory = await TradeHistory.find({ txHash: transactionHash });
       if (!existingHistory.length) {
@@ -272,16 +272,42 @@ router.post("/itemSold", service_auth, async (req, res) => {
       Logger.error("[ItemSold] Failed to save new TradeHistory: ", err.message)
     }
 
-    // remove from listing
-    await Listing.deleteMany({
-      owner: seller,
-      minter: nft,
-      tokenID: tokenId,
-      blockNumber: { $lte: blockNumber}
-    });
+    try {
+      // checks if user listens
+      const listing = await Listing.findOne({ owner: seller, minter: nft, tokenID: tokenId });
+      
+      if (quantity >= listing.quantity) {
+        // remove from listing
+        await Listing.deleteMany({
+          owner: seller,
+          minter: nft,
+          tokenID: tokenId,
+          blockNumber: { $lte: blockNumber }
+        });
+      }
+      else
+      {
+        // reduce listing
+        await Listing.updateOne({
+          owner: seller,
+          minter: nft,
+          tokenID: tokenId,
+          blockNumber: { $lte: blockNumber }
+        },{
+          $inc : {"quantity": (quantity*-1)}
+        });
+      }
+
+    } catch (err) {
+      Logger.error("[ItemSold] Failed to udpate date listing: ", err.message);
+    }
+
+
+
+
 
     Logger.info("[ItemSold] Success: ", { transactionHash, blockNumber });
-    return res.json({status: "success"});
+    return res.json({ status: "success" });
   } catch (error) {
     Logger.info("[ItemSold] Failed: ", { transactionHash: req.body.transactionHash, blockNumber: req.body.blockNumber });
     Logger.error(error);
@@ -293,7 +319,7 @@ router.post("/itemSold", service_auth, async (req, res) => {
 
 router.post("/itemUpdated", service_auth, async (req, res) => {
   try {
-    const {args, blockNumber, transactionHash } = req.body;
+    const { args, blockNumber, transactionHash } = req.body;
     const [ownerC, nftC, tokenIdBN, paytokenC, newPricePerItemBN] = args;
 
     const owner = ownerC.toLowerCase();
@@ -305,12 +331,12 @@ router.post("/itemUpdated", service_auth, async (req, res) => {
 
     // update the price of the nft here
     // first update the token price
-    let category = await Category.findOne({minterAddress: nft});
+    let category = await Category.findOne({ minterAddress: nft });
     if (category) {
       let token = await NFTITEM.findOne({
         contractAddress: nft,
         tokenID: tokenId,
-        blockNumber: {$lt: blockNumber}
+        blockNumber: { $lt: blockNumber }
       });
       if (token) {
         token.price = newPricePerItem;
@@ -324,7 +350,7 @@ router.post("/itemUpdated", service_auth, async (req, res) => {
       owner: owner,
       minter: nft,
       tokenID: tokenId,
-      blockNumber: {$lt: blockNumber},
+      blockNumber: { $lt: blockNumber },
     });
     if (list) {
       list.price = newPricePerItem;
@@ -342,7 +368,7 @@ router.post("/itemUpdated", service_auth, async (req, res) => {
     }
 
     Logger.info("[ItemUpdated] Success: ", { transactionHash, blockNumber });
-    return res.json({status: "success"});
+    return res.json({ status: "success" });
   } catch (error) {
     Logger.info("[ItemUpdated] Failed: ", { transactionHash: req.body.transactionHash, blockNumber: req.body.blockNumber });
     Logger.error(error);
@@ -353,7 +379,7 @@ router.post("/itemUpdated", service_auth, async (req, res) => {
 
 router.post("/itemCanceled", service_auth, async (req, res) => {
   try {
-    const {args, blockNumber, transactionHash} = req.body;
+    const { args, blockNumber, transactionHash } = req.body;
     const [ownerC, nftC, tokenIdBN] = args;
 
     const owner = ownerC.toLowerCase();
@@ -365,7 +391,7 @@ router.post("/itemCanceled", service_auth, async (req, res) => {
       let token = await NFTITEM.findOne({
         contractAddress: nft,
         tokenID: tokenId,
-        blockNumber: {$lt: blockNumber},
+        blockNumber: { $lt: blockNumber },
       });
       if (token) {
         token.price = 0;
@@ -380,11 +406,11 @@ router.post("/itemCanceled", service_auth, async (req, res) => {
       owner: owner,
       minter: nft,
       tokenID: tokenId,
-      blockNumber: {$lt: blockNumber}
+      blockNumber: { $lt: blockNumber }
     });
 
     Logger.info("[ItemCanceled] Success: ", { transactionHash, blockNumber });
-    return res.json({status: "success"});
+    return res.json({ status: "success" });
   } catch (error) {
     Logger.info("[ItemCanceled] Failed!: ", { transactionHash: req.body.transactionHash, blockNumber: req.body.blockNumber });
     Logger.error(error);
@@ -395,7 +421,7 @@ router.post("/itemCanceled", service_auth, async (req, res) => {
 
 router.post("/offerCreated", service_auth, async (req, res) => {
   try {
-    const {args, blockNumber, transactionHash} = req.body;
+    const { args, blockNumber, transactionHash } = req.body;
     const [creatorC, nftC, tokenIdBN, quantityBN, paytokenC, pricePerItemBN, deadlineBN] = args;
 
     const creator = creatorC.toLowerCase();
@@ -412,7 +438,7 @@ router.post("/offerCreated", service_auth, async (req, res) => {
         creator: creator,
         minter: nft,
         tokenID: tokenId,
-        blockNumber: {$lt: blockNumber},
+        blockNumber: { $lt: blockNumber },
       });
 
       const existingOffer = await Offer.find({
@@ -486,7 +512,7 @@ router.post("/offerCreated", service_auth, async (req, res) => {
     }
 
     Logger.info("[OfferCreated] Success: ", { transactionHash, blockNumber });
-    return res.json({status: "success"});
+    return res.json({ status: "success" });
   } catch (error) {
     Logger.info("[OfferCreated] Failed!: ", { transactionHash: req.body.transactionHash, blockNumber: req.body.blockNumber });
     Logger.error(error);
@@ -497,7 +523,7 @@ router.post("/offerCreated", service_auth, async (req, res) => {
 
 router.post("/offerCanceled", service_auth, async (req, res) => {
   try {
-    const {args, blockNumber, transactionHash} = req.body;
+    const { args, blockNumber, transactionHash } = req.body;
     const [creatorC, nftC, tokenIdBN] = args;
 
     const creator = creatorC.toLowerCase();
@@ -508,7 +534,7 @@ router.post("/offerCanceled", service_auth, async (req, res) => {
       creator: creator,
       minter: nft,
       tokenID: tokenId,
-      blockNumber: {$lt: blockNumber},
+      blockNumber: { $lt: blockNumber },
     });
 
     // now send email
@@ -555,7 +581,7 @@ router.post("/offerCanceled", service_auth, async (req, res) => {
               nftAddress: nft,
             };
             if (creatorAlias != alias) await sendEmail(data);
-          } 
+          }
         } else if (category == 1155) {
         }
       }
@@ -565,7 +591,7 @@ router.post("/offerCanceled", service_auth, async (req, res) => {
     }
 
     Logger.info("[OfferCanceled] Success: ", { transactionHash, blockNumber });
-    return res.json({status: "success"});
+    return res.json({ status: "success" });
   } catch (error) {
     Logger.info("[OfferCanceled] Failed!: ", { transactionHash: req.body.transactionHash, blockNumber: req.body.blockNumber });
     Logger.error(error);
